Microservices/SecretHunter/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
yara-python==4.5.2

androguard==4.1.2